correção do erro  KeyboardInterrupt

+++ src/services/fastboot_service.py
@@ -6,4 +6,11 @@ class FastbootService:
         return self.executor.run(["fastboot", "devices"])
 
     def reboot(self):
+        if not self.devices():
+            return "Erro: Nenhum dispositivo Fastboot detectado."
+        return self.executor.run(["fastboot", "reboot"])
+
+    def boot(self):
+        if not self.devices():
+            return "Erro: Nenhum dispositivo Fastboot detectado."
         return self.executor.run(["fastboot", "reboot"])
